fix(hostedFields): work in suggestion from review
