+++ src/hooks/payPalHostedFieldsHooks.ts
@@ -1,11 +1,12 @@
 import { useContext } from "react";
 
 import { PayPalHostedFieldsContext } from "../context/payPalHostedFieldsContext";
+import { contextNotEmptyValidator } from "./contextValidator";
 import type {
     PayPalHostedFieldsContextState,
     PayPalHostedFieldsAction,
 } from "../types/payPalHostedFieldTypes";
-import { contextNotEmptyValidator } from "./contextValidator";
+import type { HostedFieldsHandler } from "@paypal/paypal-js/types/components/hosted-fields";
 
 /**
  * Custom hook to get access to the PayPal Hosted Fields context
@@ -26,4 +27,10 @@ export function usePayPalHostedFieldsContext(): [
         payPalHostedFieldsContext,
         payPalHostedFieldsContext.dispatch as React.Dispatch<PayPalHostedFieldsAction>,
     ];
+}
+
+export function usePayPalHostedFields(): HostedFieldsHandler | undefined {
+    const [payPalHostedFieldsContext] = usePayPalHostedFieldsContext();
+
+    return payPalHostedFieldsContext.cardFields;
 }
